Update scripts for CHESS+ tests in GCP

+++ daml/gcp/load-runner.sh
@@ -18,10 +18,26 @@ stop_prometheus () {
   stop_network
 }
 
+default_env_variables () {
+  export DEFAULT_SPIDER_IMAGE_TAG=1.30.0
+  export DEFAULT_DAML_SDK_VERSION=1.4.0
+  export DEFAULT_MARKET_FLAVOUR=sample
+  export DEFAULT_LOAD_RUNNER_TRADE_TIMEOUT=180
+  export DEFAULT_LOAD_RUNNER_CONCURRENCY=8
+  export DEFAULT_LOAD_RUNNER_REQUESTS=333
+}
+
 initialize_env_variables () {
+  default_env_variables
   ## Define image version variables
-  export SPIDER_IMAGE_TAG=${SPIDER_IMAGE_TAG:-1.23.155}
-  export DAML_SDK_VERSION=${DAML_SDK_VERSION:-0.13.54}
+  export SPIDER_IMAGE_TAG=${SPIDER_IMAGE_TAG:-$DEFAULT_SPIDER_IMAGE_TAG}
+  export DAML_SDK_VERSION=${DAML_SDK_VERSION:-$DEFAULT_DAML_SDK_VERSION}
+  export MARKET_FLAVOUR=${MARKET_FLAVOUR:-$DEFAULT_MARKET_FLAVOUR}
+  export LOAD_RUNNER_TRADE_TIMEOUT=${LOAD_RUNNER_TRADE_TIMEOUT:-$DEFAULT_LOAD_RUNNER_TRADE_TIMEOUT}
+  export LOAD_RUNNER_CONCURRENCY=${LOAD_RUNNER_CONCURRENCY:-$DEFAULT_LOAD_RUNNER_CONCURRENCY}
+  export LOAD_RUNNER_REQUESTS=${LOAD_RUNNER_REQUESTS:-$DEFAULT_LOAD_RUNNER_REQUESTS}
+  ## Set spider app metrics port to something that doesn't clash:
+  SPIDER_APP_METRICS_PORT=54321
 }
 
 load_docker_images() {
@@ -35,7 +51,7 @@ start_concord () {
   (cd docker && ./gen-docker-concord-config.sh config-public/dockerConfigurationInput-daml.yaml)
   (cd docker && docker-compose -p "docker" --project-directory . -f ../daml/gcp/docker-compose-daml-gcp.yml up -d)
   ## Let concord stabilize
-  sleep 120
+  sleep 60
 }
 
 stop_concord () {
@@ -72,30 +88,31 @@ start_spider () {
 }
 
 stop_spider () {
-  docker kill spider-${SPIDER_IMAGE_TAG}
+  [ "$(docker ps | grep spider)" ] && docker kill spider-${SPIDER_IMAGE_TAG}
 }
 
 create_prerequisites () {
   ## Extract the necessary parties from the test data set
-  allocate-ledger-party 00001 $(_get_all_sample_parties)
+  allocate-ledger-party 00001 $(_get_all_sample_parties ${MARKET_FLAVOUR})
 
   ## Create genesis contracts
   market-setup | grep -v akka
   ## Load sample data for load-runner
-  load-sample-data | grep -v akka
+  import-data-set ${MARKET_FLAVOUR} | grep -v akka
 }
 
 run_load_runner () {
   ## Open the market
   load-runner --simulation bmw.open-market
+  sleep 10
   ## Run the benchmark tool
   load-runner \
         --simulation fix-trade.standard \
-        --trade-file /home/dlt/app/spider-load-tests/data/sample/fix_ae.tsv \
+        --trade-file /home/dlt/app/spider-load-tests/data/${MARKET_FLAVOUR}/fix_ae.tsv \
         --loop-file \
-        --trade-timeout 60s \
-        --requests 333 \
-        --concurrency 12 \
+        --trade-timeout ${LOAD_RUNNER_TRADE_TIMEOUT}s \
+        --requests ${LOAD_RUNNER_REQUESTS} \
+        --concurrency ${LOAD_RUNNER_CONCURRENCY} \
         --spec
 }
 
@@ -130,6 +147,7 @@ stop_all () {
 }
 
 print_help () {
+  default_env_variables
   echo -e "Usage: ./load-runner.sh COMMAND\n"
   echo "Available commands:"
   echo "help                     Show this help text"
@@ -138,8 +156,14 @@ print_help () {
   echo "sprom                    Stop prometheus"
   echo "stop                     Stop all containers including prometheus"
   echo ""
-  echo "To test specific spider version, set the SPIDER_IMAGE_TAG environment variable."
-  echo "To test a specific DAML SDK version, set DAML_SDK_VERSION variable."
+  echo "Modify behavior of the test through environment variables:"
+  echo "LEDGER_IMAGE      - pull a ledger docker image, default: $DEFAULT_LEDGER_IMAGE"
+  echo "SPIDER_IMAGE_TAG  - test a specific spider version, default: $DEFAULT_SPIDER_IMAGE_TAG"
+  echo "DAML_SDK_VERSION  - test a specific DAML SDK version, default: $DEFAULT_DAML_SDK_VERSION"
+  echo "MARKET_FLAVOUR    - select CHESS data flavor (sample, cde7), default: $DEFAULT_MARKET_FLAVOUR"
+  echo "LOAD_RUNNER_TRADE_TIMEOUT    - define how long load-runner should wait for asynchronous responses, default: $DEFAULT_LOAD_RUNNER_TRADE_TIMEOUT"
+  echo "LOAD_RUNNER_CONCURRENCY      - define how many threads should load-runner use for communicating with the ledger, default: $DEFAULT_LOAD_RUNNER_CONCURRENCY"
+  echo "LOAD_RUNNER_REQUESTS         - define how many trade registration requests should load-runner make, default: $DEFAULT_LOAD_RUNNER_REQUESTS"
 }
 
 OPERATION=${1:-"help"}
